Create feature to update students
Was created a feature to update students in database with the file
'update-student-name.php'.

This feature searches for a student and update it inside the database, it
gets the arguments through the command line and change it.

To update a student, in the command line should to be called the file,
informed an id and the new name: 'php
src/commands/update-student-name.php 10 "New Name"'.